docker/proxy/requirements.txt
requests
pyyaml
pytest